backend/pyproject.toml
[project]
name = "russian-laws"
version = "0.1.0"
description = "RAG сервис для поиска статей российских законов и кодексов"
authors = [
    { name = "Alex Mart", email = "alexmartynov811@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "fire>=0.7.1",
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "mlflow>=3.8.1",
    "numpy>=2.4.0",
    "openai>=2.21.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.16.2",
    "torch>=2.9.1",
    "transformers>=4.48.0",
    "uvicorn[standard]>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "dvc[s3]>=3.65.0",
    "g4f>=6.8.1",
    "ipykernel>=7.1.0",
    "pre-commit>=4.5.1",
    "scikit-learn>=1.8.0",
]

[tool.ruff]
line-length = 100
format = { docstring-code-format = true }

[tool.ruff.lint]
select = [
    # isort
    "I",
    # pycodestyle
    "E", "W",
    # Pyflakes
    "F"
]

pycodestyle = { max-doc-length = 105 }
